feat: project structure
