+++ .nvmrc
@@ -1,0 +1,1 @@
+18.17.1
