+++ package.json
@@ -3,7 +3,16 @@
   "version": "1.0.0",
   "private": false,
   "scripts": {
-    "start": "node app.js"
+    "start": "node ./src/app.js",
+    "start:prod": "NODE_ENV=production node ./src/app.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/kitamersion/kita-auth-proxy"
+  },
+  "license": "MIT",
+  "author": {
+    "name": "kitamersion"
   },
   "dependencies": {
     "axios": "^1.7.2",
@@ -15,5 +24,8 @@
     "jade": "^0.31.2",
     "morgan": "~1.10.0",
     "qs": "^6.12.1"
+  },
+  "engines": {
+    "node": "18.17.1"
   }
 }
