+++ public/stylesheets/style.css
@@ -1,3 +1,8 @@
+html {
+  background-color: #1A202C;
+  color: white
+}
+
 body {
   padding: 50px;
   font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
