--- requests/oauth/authorize.http
@@ -1,2 +1,0 @@
-POST http://localhost:3000/oauth/authorize HTTP/1.1
-content-type: application/json
