+++ src/app.js
@@ -7,7 +7,7 @@ var bodyParser = require('body-parser');
 require('dotenv').config();
 
 var indexRouter = require('./routes/index');
-var oauthRouter = require('./routes/oauth');
+var oauthRouter = require('./routes/mal/oauth');
 
 var app = express();
 
@@ -20,10 +20,10 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(bodyParser.json());
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(express.static(path.join(__dirname, '../public')));
 
 app.use('/', indexRouter);
-app.use('/oauth', oauthRouter);
+app.use('/mal/oauth', oauthRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
@@ -41,9 +41,9 @@ app.use(function(err, req, res, next) {
   res.render('error');
 });
 
-
-app.listen(3000, function() {
-  console.log('App listening on port 3000');
+const port = process.env.PORT || 3000;
+app.listen(port, function() {
+  console.log(`App listening on port http://localhost:${port}`);
 });
 
 module.exports = app;
