+++ src/requests/oauth/authorize.http
@@ -1,0 +1,2 @@
+POST http://localhost:3000/mal/oauth/authorize HTTP/1.1
+content-type: application/json
